Rename Hamming as HammingDistCalculator
